+++ riscv-rt/CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+- Limit rustc cfg flags to `riscvi`, `riscvm`, `riscvf`, and `riscvd`.
+- Temporary use of `RISCV_RT_LLVM_ARCH_PATCH` environment variable to include the
+  temporary patch required for avoid LLVM spurious errors.
+- `riscv-rt` now use the `RISCV_RT_BASE_ISA` environment variable to configure the behavior
+  of `riscv-rt-macros` depending on aspects of the base ISA (e.g., RV32I or RV32E).
+- Use `riscv-target-parser` in build script to identify target-specific configurations.
+- Add documentation to trap frame fields.
+- Avoid using `t3`+ in startup assembly to ensure compatibility with RVE.
 - `link.x.in`: remove references to `eh_frame`.
 - Rename start/end section symbols to align with `cortex-m-rt`:
     - `_stext`: it remains, as linker files can modify it.